requirements.txt
langchain >= 0.0.184
openai >= 0.27.7
fastapi>=0.95.2
langchain-openai
python-dotenv
langchain_chroma
uvicorn
langchainhub